Make everything work
The original repo has not been updated since 2014, so it was running a very early version of Kotlin, and a pretty early version of Gradle. This means that the code written for it back then didn't actually work with modern Kotlin versions, as some of the syntax was changed since.

The testing framework used has also been changed from the early version of "spek" it was using to "kotlintest".

Some files have also had their names changed.

Some general things have also been changed besides the code, like the renaming of "License.txt" to "LICENSE" and the conversion of the "build.gradle" script to "build.gradle.kts".

The actual gradle version has also been changed from 1.9 to 5.2.1, and the gradle wrapper has been updated accordingly.

A new .gitignore has also been added, as the old one didn't ignore a lot of files that should've probably been ignored, as they serve no real proper purpose in making sure that the project can be ran.

Signed-off-by: Oliver Berg <oliver@berg.moe>